Live property editor for object in example scene
This adds a new example app that features a live property editor that is dynamically generated based on the Qt properties of the selected object in the scene (currently fixed to a particular sphere). I also removed old code that attempted something similar in favor of a much more generic solution.